+++ changelog.md
@@ -1,3 +1,4 @@
 - Fixed ejecting the player out of the car
+- Fixed destroy stages of tank
 
 __NOTE THAT YOU HAVE TO USE AT LEAST FORGE 28.0.105__